pause

+++ android/TrashCam/src/com/trashcam/model/DeleteFileModelManager.java
@@ -3,8 +3,6 @@ package com.trashcam.model;
 import java.io.File;
 import java.util.HashMap;
 
-import com.trashcam.Utility;
-
 public class DeleteFileModelManager {
 	private static DeleteFileModelManager instance;
 
@@ -26,6 +24,12 @@ public class DeleteFileModelManager {
 	}
 
 	public void addFile(File new_picture, int days) {
+		if(new_picture==null){
+			throw new IllegalStateException("new_picture NOT INITED");
+		}
+		if(toBeDeleted==null){
+			throw new IllegalStateException("TOBEDELETED NOT INITED");
+		}
 		removeIfExist(new_picture);
 		toBeDeleted.put(new_picture.getAbsolutePath(), new DeleteFileModel(
 				new_picture, days));
